+++ db.tf
@@ -32,11 +32,15 @@ resource "aws_db_instance" "this" {
   backup_retention_period = var.backup_retention_period
   backup_window           = "02:00-03:00"
 
+  enabled_cloudwatch_logs_exports = local.log_exports
+
   tags = local.tags
 
   lifecycle {
     ignore_changes = [username, final_snapshot_identifier]
   }
+
+  depends_on = [aws_cloudwatch_log_group.this]
 }
 
 resource "aws_db_subnet_group" "this" {